+++ torch/_inductor/kernel/__init__.py
@@ -1,1 +1,2 @@
 from . import flex, mm, mm_common, mm_plus_mm
+from .flex_gemm import lowering as flex_gemm_lowering
